# dtl/docs/openapi.yaml
openapi: 3.0.3
info:
  title: DTL API
  description: |
    API for DTL service (quote requests, appointment requests, push, settings, admin CRUD).
    See api.md for narrative documentation.
    Admin and settings endpoints are implemented as Next.js API routes; they use in-memory or
    localStorage-backed stores (no external DB yet).
  version: 1.0.0
servers:
  - url: /api
    description: Same-origin or NEXT_PUBLIC_API_URL base
...
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      tags: [Admin – Users]
      summary: List admin users
      operationId: getUsers
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
    post:
      tags: [Admin – Users]
      summary: Create user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid body
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      description: Admin user (lib/users-store.ts, dummy-users.ts).
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [Super Admin, Admin, Staff]
        active:
          type: boolean
        canManageUsers:
          type: boolean
        lastLogin:
          type: string
          format: date-time
    UserCreate:
      type: object
      required: [name, email, role]
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [Admin, Staff]
        active:
          type: boolean
        canManageUsers:
          type: boolean
...